+++ packages/victory-core/src/victory-accessible-group/victory-accessible-group.tsx
@@ -1,5 +1,4 @@
 import React from "react";
-import PropTypes from "prop-types";
 
 export interface VictoryAccessibleGroupProps {
   desc?: string;
@@ -10,48 +9,34 @@ export interface VictoryAccessibleGroupProps {
   tabIndex?: number;
 }
 
-export class VictoryAccessibleGroup extends React.Component<VictoryAccessibleGroupProps> {
-  static displayName = "VictoryAccessibleGroup";
-  static propTypes = {
-    "aria-describedby": PropTypes.string,
-    "aria-label": PropTypes.string,
-    children: PropTypes.oneOfType([
-      PropTypes.arrayOf(PropTypes.node),
-      PropTypes.node,
-    ]),
-    className: PropTypes.string,
-    desc: PropTypes.string,
-    tabIndex: PropTypes.number,
-  };
+export const VictoryAccessibleGroup = ({
+  desc,
+  children,
+  tabIndex,
+  className = "VictoryAccessibleGroup",
+  ...props
+}: VictoryAccessibleGroupProps) => {
+  const descId =
+    desc && (props["aria-describedby"] || desc.split(" ").join("-"));
 
-  static defaultProps = {
-    className: "VictoryAccessibleGroup",
-  };
-
-  render() {
-    const { desc, children, className, tabIndex } = this.props;
-    const descId =
-      desc && (this.props["aria-describedby"] || desc.split(" ").join("-"));
-
-    return desc ? (
-      <g
-        aria-label={this.props["aria-label"]}
-        aria-describedby={descId}
-        className={className}
-        tabIndex={tabIndex}
-      >
-        <desc id={descId}>{desc}</desc>
-        {children}
-      </g>
-    ) : (
-      <g
-        aria-label={this.props["aria-label"]}
-        aria-describedby={this.props["aria-describedby"]}
-        className={className}
-        tabIndex={tabIndex}
-      >
-        {children}
-      </g>
-    );
-  }
-}
+  return desc ? (
+    <g
+      aria-label={props["aria-label"]}
+      aria-describedby={descId}
+      className={className}
+      tabIndex={tabIndex}
+    >
+      <desc id={descId}>{desc}</desc>
+      {children}
+    </g>
+  ) : (
+    <g
+      aria-label={props["aria-label"]}
+      aria-describedby={props["aria-describedby"]}
+      className={className}
+      tabIndex={tabIndex}
+    >
+      {children}
+    </g>
+  );
+};