+++ packages/victory-native/src/components/victory-boxplot.tsx
@@ -1,7 +1,10 @@
 import React from "react";
 import { Dimensions } from "react-native";
 import { G } from "react-native-svg";
-import { VictoryBoxPlot as VictoryBoxPlotBase } from "victory-box-plot/es";
+import {
+  VictoryBoxPlot as VictoryBoxPlotBase,
+  VictoryBoxPlotProps,
+} from "victory-box-plot/es";
 import { VictoryLabel } from "./victory-label";
 import { VictoryContainer } from "./victory-container";
 import { Border } from "./victory-primitives/border";
@@ -9,7 +12,7 @@ import { Whisker } from "./victory-primitives/whisker";
 import { LineSegment } from "./victory-primitives/line-segment";
 import { wrapCoreComponent } from "../helpers/wrap-core-component";
 
-export const VictoryBoxPlot = wrapCoreComponent({
+export const VictoryBoxPlot = wrapCoreComponent<VictoryBoxPlotProps>({
   Component: VictoryBoxPlotBase,
   defaultProps: {
     ...VictoryBoxPlotBase.defaultProps,