+++ packages/victory-native/src/components/victory-pie.tsx
@@ -1,13 +1,13 @@
 import React from "react";
 import { Dimensions } from "react-native";
 import { G } from "react-native-svg";
-import { VictoryPie as VictoryPieBase } from "victory-pie/es";
+import { VictoryPie as VictoryPieBase, VictoryPieProps } from "victory-pie/es";
 import { VictoryLabel } from "./victory-label";
 import { VictoryContainer } from "./victory-container";
 import { Slice } from "./victory-primitives/slice";
 import { wrapCoreComponent } from "../helpers/wrap-core-component";
 
-export const VictoryPie = wrapCoreComponent({
+export const VictoryPie = wrapCoreComponent<VictoryPieProps>({
   Component: VictoryPieBase,
   defaultProps: {
     ...VictoryPieBase.defaultProps,