+++ packages/victory-native/src/components/victory-scatter.tsx
@@ -1,13 +1,16 @@
 import React from "react";
 import { Dimensions } from "react-native";
 import { G } from "react-native-svg";
-import { VictoryScatter as VictoryScatterBase } from "victory-scatter/es";
+import {
+  VictoryScatter as VictoryScatterBase,
+  VictoryScatterProps,
+} from "victory-scatter/es";
 import { VictoryLabel } from "./victory-label";
 import { VictoryContainer } from "./victory-container";
 import { Point } from "./victory-primitives/point";
 import { wrapCoreComponent } from "../helpers/wrap-core-component";
 
-export const VictoryScatter = wrapCoreComponent({
+export const VictoryScatter = wrapCoreComponent<VictoryScatterProps>({
   Component: VictoryScatterBase,
   defaultProps: {
     ...VictoryScatterBase.defaultProps,