+++ stories/victory-histogram.stories.tsx
@@ -331,7 +331,6 @@ export const CornerRadius = () => {
 
 export const GetPath = () => {
   const verticalPathFn = (props) => {
-    // eslint-disable-next-line react/prop-types
     const { x0, x1, y0, y1 } = props;
     return `M ${x0}, ${y0}
       L ${(x1 + x0) / 2}, ${y1}
@@ -340,7 +339,6 @@ export const GetPath = () => {
   };
 
   const horizontalPathFn = (props) => {
-    // eslint-disable-next-line react/prop-types
     const { x0, x1, y0, y1 } = props;
     return `M ${x0}, ${y1}
       L ${x1}, ${(y0 + y1) / 2}