+++ packages/victory-selection-container/CHANGELOG.md
@@ -1,5 +1,17 @@
 # victory-selection-container
 
+## 36.9.1
+
+### Patch Changes
+
+- Fix victory-native component prop types ([#2785](https://github.com/FormidableLabs/victory/pull/2785))
+
+## 36.9.0
+
+### Minor Changes
+
+- Remove prop-types definitions and dependency ([#2758](https://github.com/FormidableLabs/victory/pull/2758))
+
 ## 36.8.6
 
 ## 36.8.5